+++ changelog.rst
@@ -6,6 +6,7 @@ Changelog
 
  - update to support new python-dateutil, including fix of old parsing exception
  - fix testing compatibility with Flask-SQLALchemy 2.1
+ - fix testing dependencies problem in setup
 
 0.1.22 released 2016-02-18
 ==========================