+++ changelog.rst
@@ -8,6 +8,7 @@ Changelog
  - update to support new python-dateutil, including fix of old parsing exception
  - fix testing compatibility with Flask-SQLALchemy 2.1
  - fix testing dependencies problem in setup
+ - support Python 3.4 and newer
 
 0.1.22 released 2016-02-18
 ==========================